refactor: refactor all resources

+++ src/user/user.service.ts
@@ -1,29 +1,70 @@
-import { Injectable } from '@nestjs/common';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './interfaces/user.interface';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { CreateUserDto } from './dto/create-user.dto';
+import { UpdatePasswordDto } from './dto/update-password.dto';
+import { User } from './entities/user.entity';
 
 @Injectable()
 export class UserService {
   constructor(private prisma: PrismaService) {}
 
-  async create(createUserDto: CreateUserDto) {
-    return this.prisma.user.create({ data: createUserDto });
+  async create(createUserDto: CreateUserDto): Promise<User> {
+    const createdUser = await this.prisma.user.create({ data: createUserDto });
+    return new User(createdUser);
   }
 
-  async findAll() {
-    return this.prisma.user.findMany();
+  async findAll(): Promise<User[]> {
+    const users = await this.prisma.user.findMany();
+    return users.map((user) => new User(user));
   }
 
-  async findOne(id: string) {
-    return this.prisma.user.findUnique({ where: { id } });
+  async findOne(id: string): Promise<User> {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException();
+    }
+
+    return new User(user);
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
-    return this.prisma.user.update({ where: { id }, data: updateUserDto });
+  async update(
+    id: string,
+    updatePasswordDto: UpdatePasswordDto,
+  ): Promise<User> {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException();
+    }
+
+    if (user.password !== updatePasswordDto.oldPassword) {
+      throw new ForbiddenException();
+    }
+
+    const updatedUser = await this.prisma.user.update({
+      where: { id },
+      data: {
+        password: updatePasswordDto.newPassword,
+        version: user.version + 1,
+        updatedAt: new Date(Date.now()),
+      },
+    });
+
+    return new User(updatedUser);
   }
 
-  async remove(id: string) {
-    return this.prisma.user.delete({ where: { id } });
+  async remove(id: string): Promise<void> {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException();
+    }
+
+    await this.prisma.user.delete({ where: { id } });
   }
 }
